+++ engine/client/avi/avi_win.c
@@ -251,7 +251,7 @@ qboolean AVI_ACMConvertAudio( movie_state_t *Avi )
 	return true;
 }
 
-qboolean AVI_GetVideoInfo( movie_state_t *Avi, long *xres, long *yres, float *duration )
+qboolean AVI_GetVideoInfo( movie_state_t *Avi, int *xres, int *yres, float *duration )
 {
 	if( !Avi->active )
 		return false;
@@ -269,7 +269,7 @@ qboolean AVI_GetVideoInfo( movie_state_t *Avi, long *xres, long *yres, float *du
 }
 
 // returns a unique frame identifier
-long AVI_GetVideoFrameNumber( movie_state_t *Avi, float time )
+int AVI_GetVideoFrameNumber( movie_state_t *Avi, float time )
 {
 	if( !Avi->active )
 		return 0;
@@ -277,7 +277,7 @@ long AVI_GetVideoFrameNumber( movie_state_t *Avi, float time )
 	return (time * Avi->video_fps);
 }
 
-long AVI_GetVideoFrameCount( movie_state_t *Avi )
+int AVI_GetVideoFrameCount( movie_state_t *Avi )
 {
 	if( !Avi->active )
 		return 0;
@@ -285,7 +285,7 @@ long AVI_GetVideoFrameCount( movie_state_t *Avi )
 	return Avi->video_frames;
 }
 
-long AVI_TimeToSoundPosition( movie_state_t *Avi, long time )
+int AVI_TimeToSoundPosition( movie_state_t *Avi, int time )
 {
 	if( !Avi->active || !Avi->audio_stream )
 		return 0;
@@ -382,7 +382,7 @@ qboolean AVI_SeekPosition( movie_state_t *Avi, dword offset )
 }
 
 // get a chunk of audio from the stream (in bytes)
-long AVI_GetAudioChunk( movie_state_t *Avi, char *audiodata, long offset, long length )
+int AVI_GetAudioChunk( movie_state_t *Avi, char *audiodata, int offset, int length )
 {
 	long	result = 0;
 	int	i;
